part 1 solution + some part2 investigation

+++ src/main/scala/day11/binaryOp.scala
@@ -3,16 +3,16 @@ package day11
 object BinaryOperation {
   def build(opName: String, param1: String, param2: String) = {
     val paramFns = List(param1, param2).map(_ match {
-      case "old" => (x: Int) => x
+      case "old" => (x: Long) => x
       case v => {
-        val va = v.toInt
-        (x: Int) => va
+        val va = v.toLong
+        (x: Long) => va
       }
     })
 
     opName match {
-      case "+" => (x: Int) => paramFns(0)(x) + paramFns(1)(x)
-      case "*" => (x: Int) => paramFns(0)(x) * paramFns(1)(x)
+      case "+" => (x: Long) => paramFns(0)(x) + paramFns(1)(x)
+      case "*" => (x: Long) => paramFns(0)(x) * paramFns(1)(x)
     }
   }
 }
